+++ emspayafterpay/emspayafterpay.php
@@ -311,7 +311,7 @@ class emspayAfterpay extends PaymentModule
         $webhookUrl = EmspayHelper::getWebHookUrl();
  
         try {
-            $response = $this->ginger->createOrder([
+            $response = $this->ginger->createOrder(array_filter([
 		    'amount' => $totalInCents,                                                      // Amount in cents
 		    'currency' => $currency,                                                        // Currency
 		    'transactions' => [
@@ -326,7 +326,7 @@ class emspayAfterpay extends PaymentModule
 		    'extra' => ['plugin' => EmspayHelper::getPluginVersionText($this->version)],   	// Extra information
 		    'webhook_url' => $webhookUrl,                                                   // Webhook URL
 		    'order_lines' => $orderLines                                                    // Order lines
-            ]);
+            ]));
         } catch (\Exception $exception) {
             return Tools::displayError($exception->getMessage());
         }