+++ emspaybanktransfer/emspaybanktransfer.php
@@ -190,7 +190,7 @@ class emspayBanktransfer extends PaymentModule
         $webhookUrl = EmspayHelper::getWebHookUrl();
 
         try {
-            $response = $this->ginger->createOrder([
+            $response = $this->ginger->createOrder(array_filter([
 			'amount' => $totalInCents,                                                      // Amount in cents
 			'currency' => $currency,                                                        // Currency
 			'transactions' => [
@@ -204,7 +204,7 @@ class emspayBanktransfer extends PaymentModule
 			'customer' => $customer,                                                        // Customer information
 			'extra' => ['plugin' => EmspayHelper::getPluginVersionText($this->version)],    // Extra information
 			'webhook_url' => $webhookUrl,                                                   // Webhook URL
-            ]);
+            ]));
         } catch (\Exception $exception) {
             return Tools::displayError($exception->getMessage());
         }