+++ emspaywechat/emspaywechat.php
@@ -155,7 +155,7 @@ class emspayWechat extends PaymentModule
         $returnURL = $this->getReturnURL($cart);
 
         try {
-            $response = $this->ginger->createOrder([
+            $response = $this->ginger->createOrder(array_filter([
                 'amount' => $totalInCents,                                                      // Amount in cents
                 'currency' => $currency,                                                        // Currency
                 'transactions' => [
@@ -169,7 +169,7 @@ class emspayWechat extends PaymentModule
                 'customer' => $customer,                                                        // Customer information
                 'extra' => ['plugin' => EmspayHelper::getPluginVersionText($this->version)],    // Extra information
                 'webhook_url' => $webhookUrl                                                    // Webhook URL
-            ]);
+            ]));
         } catch (\Exception $exception) {
             return Tools::displayError($exception->getMessage());
         }